Rename Chronology and ScrollableText files to more clear filenames
