+++ frontend/src/components/AnimeDetailsTabs/ChronologySeason.tsx
