+++ frontend/src/components/AnimeDetailsTabs/ScrollableDescription.tsx
